move createApproveTransactionData to utils
